Bllossom-llama-3.2-ko/requirements.txt
python-dotenv
numpy
pandas
openai
torch
bitsandbytes
mlflow
#deepspeed
mpi4py
optuna